requirements.txt
python-dotenv==1.0.0
pandas==2.2.3
python-pptx==0.6.18
openpyxl==3.1.2
openai==0.28
tqdm==4.66.1
